This commit comprises of multiple changes 1. Final MotherPCB (Redesigned by Akshay, only pushed to GitHub by me) 2. PCB for interfacing with the Ethernet MagJack 3. Code for Pressure Transducer data acquisition 4. Inclusion of an ADS1263 Library modified to allow use of SPI1 5. Code for Tachometer re-written as class with static member functions

+++ Microcontroller Code/Guru/include/DataStructs.h
@@ -19,10 +19,10 @@ struct Gyan {
     uint32_t thrust; //id - 0xB0
     uint8_t currFlow; //id - 0xC0
     uint8_t setFlow;
-    uint16_t pressure[4]; //id - 0xD0
+    int32_t pressure[4]; //id - 0xD0
     uint32_t temperature[9]; //id - 0xE0
     uint8_t status;
-    // Total 46 bytes
+    // Total 54 bytes
 };
 
 struct Directive {
